Formula: Add imageValue(image)
