+++ core/formula/src/main/resources/META-INF/services/org.csstudio.apputil.formula.spi.FormulaFunction
@@ -27,11 +27,12 @@ org.csstudio.apputil.formula.math.Pow
 # Other functions
 org.csstudio.apputil.formula.demo.Factorial
 org.csstudio.apputil.formula.areadetector.ADDataTypeMappingFunction
-org.csstudio.apputil.formula.areadetector.ImageWidthFunction
 org.csstudio.apputil.formula.areadetector.ImageHeightFunction
+org.csstudio.apputil.formula.areadetector.ImageValueFunction
+org.csstudio.apputil.formula.areadetector.ImageWidthFunction
 org.csstudio.apputil.formula.areadetector.ImageXOffsetFunction
-org.csstudio.apputil.formula.areadetector.ImageYOffsetFunction
 org.csstudio.apputil.formula.areadetector.ImageXReversedFunction
+org.csstudio.apputil.formula.areadetector.ImageYOffsetFunction
 org.csstudio.apputil.formula.areadetector.ImageYReversedFunction
 
 # String
